Extract toolbar filtering and sorting
Before, filtering and sorting was tied to VM entities. After this patch
the generic filter/sort component can be reused for other entities.

Filter component changes:
1. create a new type of filter - a date filter
2. generalize name filter to text based filter - as before only one
   filter of this type is supported per filter toolbar
3. extract enum based filter as SelectFilter

Sort component changes:
1. refactor SortFields from constant to parameter
2. drop unused isNumeric flag - natural sort order is used for all
   types of input
